Adding dropdown and hide prompt
